Created SceneStateManager and UIManager Scripts
SceneStateManager handles which scene is loaded, and whether or not the game world is paused.
UIManager handles which UI Canvas Group is currently active, so we can navigate between the pause menu, how to play menu, and normal gameplay.

To test this, use P to pause and the mouse to click the UI buttons.