Added Visual Prompts
There is now a component called "Display Message" that you can attach to any game object you want the Player to interact with. It has a field where you can input custom text. When the player triggers the collider of this object, the message will be shown above the player's head. Still fleshing it out to be turned off in certain contexts.